Jarvis/python-agent/requirements.txt
browser-use
fastapi
uvicorn
langchain-google-genai
python-dotenv
playwright
pydantic